#5 Update Markup, Rename IdeaList -Divide component
