+++ src/components/idea/IdeaSection.css
@@ -1,0 +1,214 @@
+.main-idea-list-wrap {
+	background-color: #f5f5f5;
+	display: flex;
+	justify-content: center;
+}
+
+.main-idea-contents-wrap {
+	width: 1200px;
+	min-height: 538px;
+}
+
+.card {
+	background-color: #fff;
+	background-clip: border-box;
+	border: 0 rgba(0, 0, 0, 0);
+	border-radius: 0;
+}
+
+.card-header {
+	margin-bottom: 0;
+	border-bottom: 0;
+	padding: 0 0 65px 0;
+	background-color: #F5F5F5;
+}
+
+.header-left {
+	display: inline-block;
+	float: left;
+	width: 50%;
+	text-align: left;
+	margin-top: 38px;
+}
+
+.tag-result {
+	width: 248px;
+	height: 24px;
+	font-family: NotoSansCJKkr;
+	font-size: 16px;
+	font-weight: normal;
+	font-style: normal;
+	font-stretch: normal;
+	line-height: normal;
+	letter-spacing: -0.6px;
+	color: #4a4a4a;
+}
+
+.header-right {
+	display: inline-block;
+	float: right;
+	width: 50%;
+	margin-top: 25px;
+}
+
+.list-info {
+	display: inline;
+	width: 130px;
+	height: 24px;
+	font-family: NotoSansCJKkr;
+	font-size: 16px;
+	font-weight: normal;
+	font-style: normal;
+	font-stretch: normal;
+	line-height: normal;
+	letter-spacing: -0.6px;
+	color: #4a4a4a;
+	margin-right: 0.6%;
+}
+
+.on {
+	display: inline;
+	width: 102px;
+	height: 24px;
+	font-family: NotoSansCJKkr;
+	font-size: 16px;
+	font-weight: 500;
+	font-style: normal;
+	font-stretch: normal;
+	line-height: normal;
+	letter-spacing: -0.6px;
+	color: #4a4a4a;
+	margin-left: 1.3%;
+}
+
+.search {
+	display: inline-block;
+	float: right;
+}
+
+.search-input {
+	/*
+		type=search로 했을 경우, icon 넣을 수 있는 방법
+	*/
+
+	/*background-image : url('../../../assets/img/ico-search@2x.png');*/
+	/*background-position:top right;*/
+	/*background-repeat:no-repeat;*/
+
+	/*background-width: 28px;*/
+	/*background-height: 28px;*/
+	/*object-fit: contain;*/
+
+	font-family: NotoSansCJKkr;
+	font-size: 16px;
+	font-weight: 300;
+	font-style: normal;
+	font-stretch: normal;
+	line-height: normal;
+	letter-spacing: -0.53px;
+	color: #9b9b9b;
+
+	padding: 14px 0 12px 18px;
+}
+
+/*
+  검색창 type을 div로 했을 경우
+*/
+/*[contenteditable=true]:empty:before{*/
+/*content: attr(placeholder);*/
+/*}*/
+
+/*div[contenteditable=true] {*/
+/*position: relative;*/
+/*text-align: left;*/
+/*vertical-align: center;*/
+/*!*font-family: NotoSansCJKkr;*!*/
+/*!*font-size: 16px;*!*/
+/*!*font-weight: 300;*!*/
+/*!*font-style: normal;*!*/
+/*!*font-stretch: normal;*!*/
+/*!*line-height: normal;*!*/
+/*!*letter-spacing: -0.53px;*!*/
+/*!*color: #9b9b9b;*!*/
+/*}*/
+
+.Rectangle-Copy {
+	width: 103px;
+	height: 46px;
+	border-radius: 6px;
+	background-color: #273ea5;
+
+	padding: 11px 24px;
+	margin-right: 16px;
+}
+
+span {
+	width: 55px;
+	height: 24px;
+	font-family: NotoSansCJKkr;
+	font-size: 16px;
+	font-weight: normal;
+	font-style: normal;
+	font-stretch: normal;
+	line-height: normal;
+	letter-spacing: -1px;
+	color: #ffffff;
+}
+
+.Rectangle {
+	width: 333px;
+	height: 50px;
+	border-radius: 6px;
+	border: solid 1px #eeeeee;
+	background-color: #ffffff;
+}
+
+.ico_search {
+	/*float: right;*/
+	position: absolute;
+	left: 291px;
+	bottom: 11px;
+
+	width: 28px;
+	height: 28px;
+	object-fit: contain;
+}
+
+.card-body {
+	background-color: #f5f5f5;
+	padding: 0;
+}
+
+
+.titles {
+	height: 20px;
+	text-align: left;
+	/*margin-top: 65p;*/
+	margin-bottom: 17px;
+}
+
+.title {
+	display: inline-block;
+	/*height: 20px;*/
+	font-family: NotoSansCJKkr;
+	font-size: 14px;
+	font-weight: normal;
+	font-style: normal;
+	font-stretch: normal;
+	letter-spacing: -1px;
+	color: #4a4a4a;
+}
+
+.ico_table_tag {
+	width: 16px;
+	height: 16px;
+	object-fit: contain;
+	margin-left: 6px;
+}
+
+.Group-10 {
+	width: 16px;
+	height: 16px;
+	object-fit: contain;
+	margin-left: 6px;
+}
